# [PERF] account: Speed-up duplicate bills lookup in list view