Description:
------------
Following commit 4e7b5e44b292dc41217e4c50048b8bb8408246f5, the list view for bills now shows a colored highlight on the bill's reference if there is a potential duplicate, and the matching on the invoice date was relaxed to include potential bills from the same year.

This caused a performance regression, as the index on `invoice_date` could no longer be used since the year is now extracted from it.

This patch adds a partial 'btree' index on the `ref` field specifically for bills, as it is the most discriminatory factor when identifying potential duplicates.

Benchmark:
----------
On a database with 6M `account_move` records, 230k of which are `'in_invoice'` & `'in_refund'`, `_fetch_duplicate_reference`, triggered when opening the list view of bills, took:

| Before | After | Speedup |
|--------|-------|---------|
| 19.8s  | 20ms  | 990x    |


---
I confirm I have signed the CLA and read the PR guidelines at www.odoo.com/submit-pr