@RicGR98 2025-03-21
> @RicGR98 @pivi-odoo I am not sure why this year thing or the invoice date is in the ref anyways. I would think the ref should contain the entire identification, but I remember @tsb-odoo once saying that people mess around with it. But also functionally, if you put an invoice the 31st of December instead of 1st of January, you won't see the duplicate. To me, the index should above all have the commercial_partner_id, which it does not. (but why the huge speedup then?)

The year thing is important because some reset their sequence every year, so there might be multiple bills with the same ref but on different years :)

I'm not a big big fan of escaping when one could have `"` and `'` but it's subjective so IMO the PR LGTM, thanks a lot, I hadn't thought about the perfs ! I'll let you double-check Josse :)