@pivi-odoo 2025-03-21
> @RicGR98 @pivi-odoo I am not sure why this year thing or the invoice date is in the ref anyways. I would think the ref should contain the entire identification, but I remember @tsb-odoo once saying that people mess around with it. But also functionally, if you put an invoice the 31st of December instead of 1st of January, you won't see the duplicate. To me, the index should above all have the commercial_partner_id, which it does not. (but why the huge speedup then?)

@jco-odoo The speed up comes from the fact the plan changed from a non-selective Bitmap Index Scans on `company_id` AND `move_type` indexes, which churned through 40 GiB of data 🫠, to just an Index Scan on this new index, which is strongly selective (only 5.6 MiB of data touched). So it's an IO bottleneck mostly.
The most overfit index for this specific query would be based on `(ref, commercial_partner_id, company_id, move_type)`.
The order I'm showing is from most selective to least selective. This was inferred from Postgres' `pg_stats.n_distinct` catalog, `ref` has 2x more distinct values than `commercial_partner_id`, therefor is more selective. I also confirmed this by testing the same index with permutations of the keys of the index, it's fastest in the order I've mentioned.

Since the index was kinda fat, I just removed the keys that were non-selective, and keeping `commercial_partner_id` yielded no significant performance improvement (1ms vs 3ms execution time), here the planning takes the most time (16ms), so I dropped it. 

Here are the plans:
- Before: https://explain.dalibo.com/plan/2eee67ee5c12c85b
- After: https://explain.dalibo.com/plan/37a8872cg1552edf

> I'm not a big big fan of escaping when one could have " and ' but it's subjective so IMO the PR LGTM, thanks a lot, I hadn't thought about the perfs ! I'll let you double-check Josse :)

I heard you guys use `"..."` primarily for user-facing strings / translated messages, and `'...'` for technical strings. Since this is a technical string, I just did the necessary escaping.
I'm totally impartial on the usage `"..."` vs `'...'`. Let me know what you prefer.