@jco-odoo 2025-03-21
@pivi-odoo Thanks for explaining.  Ref has a trigram search, but that is not ideal for exact matching, so a separate one with bigram just for duplicate checking makes sense.  Let me know for the style stuff.  Otherwise lgtm.